[FE] RENAME: social account link modal 위치 변경 #1739
